requirements.txt
Flask==2.1.2
pyrebase4==4.6.0
